bluedevices/requirements.txt
bluepy==1.3.0
miflora==0.7.2
paho-mqtt==1.6.1